gitconfig: Update refs by default when rebasing
Splitting a large feature into multiple commits and PRs, creating a
story, creates something that's called stacked branches/PRs.

If there is a need to fix something at the bottom of the stack, rebasing
every other branch on top of it again can be a tedious, manual process.
Using the --update-refs flag during rebasing does this automatically.

Make this the default behavior when running a rebase procedure.

Signed-off-by: Dimitris Poulopoulos <dimitris.a.poulopoulos@gmail.com>